--- client/pages/Customers/CustomerDetails/CustomerActions/CustomerActions.module.scss
@@ -1,11 +1,0 @@
-.customerActions {
-  display: flex;
-  flex-direction: row;
-  padding: 0.5rem 0;
-
-  .container {
-    display: flex;
-    flex-direction: row;
-    gap: 0.5rem;
-  }
-}
