--- client/pages/Customers/CustomerDetails/CustomerActions/CustomerActions.tsx
@@ -1,48 +1,0 @@
-import { DynamicButton } from 'components'
-import { usePermissions } from 'hooks'
-import React from 'react'
-import { useTranslation } from 'react-i18next'
-import { PermissionScope as $ } from 'security'
-import { StyledComponent } from 'types'
-import { useCustomersContext } from '../../context'
-import { OPEN_CUSTOMER_PANEL } from '../../reducer/actions'
-import { useCustomerDeleteAction } from './DeleteAction'
-import styles from './CustomerActions.module.scss'
-
-/**
- * Component for displaying customer action buttons like delete and edit
- * 
- * @category Customers
- */
-export const CustomerActions: StyledComponent = (props) => {
-  const { state, dispatch } = useCustomersContext()
-  const [, hasPermission] = usePermissions()
-  const { t } = useTranslation()
-  const onDelete = useCustomerDeleteAction()
-
-  return (
-    <div className={styles.customerActions} hidden={props.hidden}>
-      <div className={styles.container}>
-        <DynamicButton
-          hidden={!hasPermission($.DELETE_CUSTOMERS)}
-          text={t('customers.deleteButtonLabel')}
-          iconName='Delete'
-          {...onDelete}
-          disabled={!Boolean(state.selected)}
-          transparent
-        />
-        <DynamicButton
-          hidden={!hasPermission($.MANAGE_CUSTOMERS)}
-          text={t('customers.editButtonLabel')}
-          iconName='Edit'
-          onClick={() => dispatch(OPEN_CUSTOMER_PANEL({ onDismissCallback: () => {} }))}
-          disabled={!Boolean(state.selected)}
-          transparent
-        />
-      </div>
-      {onDelete.dialog}
-    </div>
-  )
-}
-
-CustomerActions.displayName = 'CustomerActions'
