+++ client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/DeleteAction/CustomerDeleteDialog.tsx
