+++ client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/DeleteAction/deleteCustomer.gql
