+++ client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/DeleteAction/index.ts
