+++ client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/DeleteAction/projects.gql
