+++ client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/DeleteAction/types.ts
