+++ client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/DeleteAction/useCustomerDeleteAction.tsx
@@ -4,8 +4,8 @@ import React, { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useHistory } from 'react-router-dom'
 import { BaseResult, Customer, Project } from 'types'
-import { useCustomersContext } from '../../../context'
-import { CUSTOMER_DELETE_SUCCESS } from '../../../reducer'
+import { useCustomersContext } from '../../../../context'
+import { CUSTOMER_DELETE_SUCCESS } from '../../../../reducer'
 import $deleteCustomer from './deleteCustomer.gql'
 import { CustomerDeleteDialog } from './CustomerDeleteDialog'
 import $projectsQuery from './projects.gql'
