--- client/pages/Customers/CustomerDetails/CustomerHeader/CustomerActions/index.tsx
@@ -1,1 +1,0 @@
-export * from './CustomerActions'
